Convert from React Router to Vite + React setup
