--- app/routes.ts
@@ -1,5 +1,0 @@
-import { type RouteConfig, index, route } from "@react-router/dev/routes";
-
-export default [
-    index("routes/_index.tsx"),
-] satisfies RouteConfig;
