+++ src/components/Header.tsx
@@ -1,15 +1,20 @@
-import { Link } from "react-router";
-
 export function Header() {
+    const scrollToTop = () => {
+        window.scrollTo({ top: 0, behavior: 'smooth' });
+    };
+
     return (
         <div className="flex justify-center sticky top-0 z-50 bg-gray-900">
             <header className="w-7/10 mt-6 mb-2">
                 <div className="px-6 py-4">
                     <div className="grid grid-cols-12 items-center">
                         <div className="col-span-6">
-                            <Link to="/portfolio/" className="text-3xl font-medium text-white hover:text-blue-300 transition-colors">
+                            <button 
+                                onClick={scrollToTop}
+                                className="text-3xl font-medium text-white hover:text-blue-300 transition-colors bg-transparent border-none cursor-pointer"
+                            >
                                 Andy Wang
-                            </Link>
+                            </button>
                         </div>
                             
                         <div className="col-span-6 flex items-center justify-end gap-3">
