+++ src/components/MiscSection.tsx
