+++ vite.config.ts
@@ -1,9 +1,13 @@
-import { reactRouter } from "@react-router/dev/vite";
-import tailwindcss from "@tailwindcss/vite";
-import { defineConfig } from "vite";
-import tsconfigPaths from "vite-tsconfig-paths";
+import { defineConfig } from 'vite'
+import react from '@vitejs/plugin-react'
+import tailwindcss from '@tailwindcss/vite'
 
 export default defineConfig({
-  plugins: [tailwindcss(), reactRouter(), tsconfigPaths()],
-  base: "/portfolio/",
-});
+  plugins: [react(), tailwindcss()],
+  base: '/portfolio/',
+  resolve: {
+    alias: {
+      '@': '/src',
+    },
+  },
+})
